fix: unecessary rerenders
